centralization/requirements.txt
fastapi
uvicorn[standart]
asyncpg
pydantic
python-dotenv
